+++ crates/compiler/tests/fixtures/partial_bundling/basic/output.js
@@ -1,5 +1,5 @@
-//__farm_runtime.450b140c.mjs:
- (globalThis || window || self || global)['__farm_default_namespace__'] = {__FARM_TARGET_ENV__: 'browser'};(function (modules, entryModule) {
+//__farm_runtime.67cba0a2.mjs:
+ (globalThis || window || global)['__farm_default_namespace__'] = {__FARM_TARGET_ENV__: 'browser'};(function (modules, entryModule) {
             var cache = {};
 
             function dynamicRequire(id) {
@@ -21,8 +21,7 @@
           
             require(entryModule);
           })({"d2214aaa": function(module, exports, farmRequire, farmDynamicRequire) {
-"use strict";
-console.log("runtime/index.js")(globalThis || window || self || global)["__farm_default_namespace__"].__farm_module_system__.setPlugins([]);
+console.log("runtime/index.js")(globalThis || window || global)["__farm_default_namespace__"].__farm_module_system__.setPlugins([]);
 
 },}, "d2214aaa");
 
@@ -30,7 +29,7 @@ console.log("runtime/index.js")(globalThis || window || self || global)["__farm_
  (function (modules) {
             for (var key in modules) {
               modules[key].__farm_resource_pot__ = 'debounce_6f74.js';
-                (globalThis || window || self || global)['__farm_default_namespace__'].__farm_module_system__.register(key, modules[key]);
+                (globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.register(key, modules[key]);
             }
         })({"01609b59": function(module, exports, farmRequire, farmDynamicRequire) {
 "use strict";
@@ -52,10 +51,10 @@ function debounce(fn) {
 },});
 
 //index.js:
- import "./__farm_runtime.450b140c.mjs";import "./index_2faa.js";import "./index_64d2.js";(globalThis || window || self || global)['__farm_default_namespace__'].__farm_module_system__.setExternalModules({ "module": { ...((globalThis || window || self || {})['module'] || {}), __esModule: true } });(function (modules) {
+ import "./__farm_runtime.67cba0a2.mjs";import "./index_2faa.js";import "./index_64d2.js";(globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.setExternalModules({ "module": { ...((globalThis || window || {})['module'] || {}), __esModule: true } });(function (modules) {
             for (var key in modules) {
               modules[key].__farm_resource_pot__ = 'index_dcdc.js';
-                (globalThis || window || self || global)['__farm_default_namespace__'].__farm_module_system__.register(key, modules[key]);
+                (globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.register(key, modules[key]);
             }
         })({"b5d64806": function(module, exports, farmRequire, farmDynamicRequire) {
 "use strict";
@@ -87,13 +86,13 @@ farmDynamicRequire("01609b59").then((debounce)=>{
 });
 var _default = defineConfig({});
 
-},});(globalThis || window || self || global)['__farm_default_namespace__'].__farm_module_system__.setInitialLoadedResources(['index_2faa.js','index_64d2.js']);(globalThis || window || self || global)['__farm_default_namespace__'].__farm_module_system__.setDynamicModuleResourcesMap({ '01609b59': [{ path: 'debounce_6f74.js', type: 'script' },{ path: 'index_2faa.js', type: 'script' },] });var farmModuleSystem = (globalThis || window || self || global)['__farm_default_namespace__'].__farm_module_system__;farmModuleSystem.bootstrap();var entry = farmModuleSystem.require("b5d64806");export default entry.default || entry;
+},});(globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.setInitialLoadedResources(['index_2faa.js','index_64d2.js']);(globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.setDynamicModuleResourcesMap({ '01609b59': [{ path: 'debounce_6f74.js', type: 'script' },{ path: 'index_2faa.js', type: 'script' },] });var farmModuleSystem = (globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__;farmModuleSystem.bootstrap();var entry = farmModuleSystem.require("b5d64806");export default entry.default || entry;
 
 //index_2faa.js:
  (function (modules) {
             for (var key in modules) {
               modules[key].__farm_resource_pot__ = 'index_2faa.js';
-                (globalThis || window || self || global)['__farm_default_namespace__'].__farm_module_system__.register(key, modules[key]);
+                (globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.register(key, modules[key]);
             }
         })({"a5831d05": function(module, exports, farmRequire, farmDynamicRequire) {
 "use strict";
@@ -116,7 +115,7 @@ function debug(msg) {
  (function (modules) {
             for (var key in modules) {
               modules[key].__farm_resource_pot__ = 'index_64d2.js';
-                (globalThis || window || self || global)['__farm_default_namespace__'].__farm_module_system__.register(key, modules[key]);
+                (globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.register(key, modules[key]);
             }
         })({"726cd210": function(module, exports, farmRequire, farmDynamicRequire) {
 "use strict";