+++ crates/plugin_html/tests/mod.rs
@@ -66,10 +66,16 @@ fn html_build_stage() {
 
     assert_eq!(
       analyze_deps_param.deps,
-      vec![PluginAnalyzeDepsHookResultEntry {
-        source: "./src/main.ts".to_string(),
-        kind: ResolveKind::ScriptSrc
-      }]
+      vec![
+        PluginAnalyzeDepsHookResultEntry {
+          source: "/favicon.ico".to_string(),
+          kind: ResolveKind::LinkHref
+        },
+        PluginAnalyzeDepsHookResultEntry {
+          source: "./src/main.ts".to_string(),
+          kind: ResolveKind::ScriptSrc
+        }
+      ]
     );
   });
 }