+++ examples/react/src/main.tsx
@@ -4,8 +4,11 @@ import './main.css';
 import { Button } from '@farmfe-examples/lib-for-browser';
 
 export function Main() {
+  console.log(1231);
+  
   return (
     <>
+    123213
       <Button />
       <Welcome />
     </>